pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.12",
    "flake8>=6.0",
    "mypy>=1.0",
    "ruff>=0.13.1",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "profitability-checker"
version = "0.1.0"
description = "Tellor Layer Profitability Checker for Cosmos SDK oracle chains"
readme = "README.md"
requires-python = ">=3.8.1"
dependencies = [
    "pyyaml>=6.0",
    "termcolor>=2.3",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
]


[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
target-version = "py38"
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[project.scripts]
profitability-checker = "src.checker:main"
prof-check = "src.checker:main"
